+++ android/app/src/main/java/de/railslove/meinecheckins/MainActivity.java
@@ -1,4 +1,4 @@
-package de.railslove.wfdcheckin;
+package de.railslove.meinecheckins;
 
 import com.facebook.react.ReactActivity;
